Use own file for user meta.
