+++ meom-dodo.php
@@ -24,3 +24,4 @@ add_action( 'plugins_loaded', 'meom_dodo_i18n', 2 );
 require_once dirname( __FILE__ ) . '/inc/contact-page.php';
 require_once dirname( __FILE__ ) . '/inc/enqueue-assets.php';
 require_once dirname( __FILE__ ) . '/inc/remove-editor-settings.php';
+require_once dirname( __FILE__ ) . '/inc/user-meta.php';
